Begin refactoring the transformers loader
